apps/face-service/pyproject.toml
[project]
name = "face-service"
version = "0.1.0"
description = "Attndly Face Recognition Microservice"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "pydantic-settings>=2.7.0",
    "httpx>=0.28.0",
    "deepface>=0.0.93",
    "opencv-python-headless>=4.9.0",
    "numpy>=1.26.0",
    "pillow>=10.0.0",
    "python-multipart>=0.0.9",
]

[dependency-groups]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.25.0",
    "httpx>=0.28.0",
    "ruff>=0.9.0",
    "mypy>=1.14.0",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
